requirements.txt
# Hackathon evaluation system
langgraph>=0.2.55
langchain>=0.3.14
langchain-openai>=0.2.14
langchain-community>=0.3.14
langchain-experimental>=0.3.4
gitpython>=3.1.43
pydantic>=2.9.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
faiss-cpu>=1.8.0
langchain-text-splitters>=0.3.4
typing_extensions>=4.12.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
PyGithub>=2.4.0
pypdf>=5.0.0
pdfplumber>=0.11.0
openpyxl>=3.1.0
